+++ Paradigm/src/utopia/paradigm/measurement/DistanceUnit.scala
@@ -1,5 +1,7 @@
 package utopia.paradigm.measurement
 
+import MetricScale._
+
 /**
  * An enumeration for standard distance units
  * @author Mikko Hilpinen
@@ -20,7 +22,7 @@ sealed trait DistanceUnit
 	 * @param targetUnit Another unit
 	 * @return A modifier that must be applied to a number when converting from this unit to the other unit
 	 */
-	def conversionModifierFor(targetUnit: DistanceUnit): Double
+	def conversionModifierTo(targetUnit: DistanceUnit): Double
 	
 	
 	// COMPUTED -----------------------------
@@ -34,99 +36,191 @@ sealed trait DistanceUnit
 	/**
 	 * @return A modifier from this unit to millimeters
 	 */
-	def toMm = conversionModifierFor(MilliMeter)
+	def toMm = conversionModifierTo(MilliMeter)
 	/**
 	 * @return a modifier from this unit to centimeters
 	 */
-	def toCm = conversionModifierFor(CentiMeter)
+	def toCm = conversionModifierTo(CentiMeter)
 	/**
 	 * @return A modifier from this unit to meters
 	 */
-	def toM = conversionModifierFor(Meter)
+	def toM = conversionModifierTo(Meter)
 	/**
 	 * @return A modifier from this unit to kilometers
 	 */
-	def toKm = conversionModifierFor(KiloMeter)
+	def toKm = conversionModifierTo(KiloMeter)
 	
 	/**
 	 * @return A modifier from this unit to inches
 	 */
-	def toInch = conversionModifierFor(Inch)
+	def toInch = conversionModifierTo(Inch)
 	/**
 	 * @return A modifier from this unit to feet
 	 */
-	def toFeet = conversionModifierFor(Feet)
+	def toFeet = conversionModifierTo(Feet)
 	
 	
 	// IMPLEMENTED  ---------------------
 	
 	override def toString = abbreviation
+	
+	
+	// OTHER    -------------------------
+	
+	/**
+	  * @param originUnit A unit
+	  * @return Modifier that needs to be applied to a value in the specified unit
+	  *         in order to receive a value in this unit
+	  */
+	def conversionModifierFrom(originUnit: DistanceUnit) = 1.0 / conversionModifierTo(originUnit)
+	
+	@deprecated("Please use .conversionModifierTo(DistanceUnit) or .conversionModifierFrom(DistanceUnit) instead", "1.5")
+	def conversionModifierFor(targetUnit: DistanceUnit): Double = conversionModifierTo(targetUnit)
 }
 
 object DistanceUnit
 {
 	// TODO: Add a graph-based algorithm for unit conversions
 	
+	// NESTED   -------------------------
+	
+	object MeterUnit extends MetricUnitFactory[MeterUnit]
+	{
+		// ATTRIBUTES   -----------------
+		
+		private lazy val scaleMap = Map[MetricScale, MeterUnit](
+			Default -> Meter,
+			Deci -> DeciMeter,
+			Centi -> CentiMeter,
+			Milli -> MilliMeter,
+			Micro -> MicroMeter,
+			Kilo -> KiloMeter,
+			Mega -> MegaMeter
+		)
+		
+		
+		// IMPLEMENTED  -----------------
+		
+		override def apply(scale: MetricScale): MeterUnit = scaleMap(scale)
+	}
+	
+	sealed trait MeterUnit extends DistanceUnit with MetricUnit[MeterUnit]
+	{
+		// IMPLEMENTED  -----------------
+		
+		override def self: MeterUnit = this
+		override def abbreviation: String = s"${scale.prefix}m"
+		
+		override protected def factory: MetricUnitFactory[MeterUnit] = MeterUnit
+		
+		
+		// OTHER    ---------------------
+		
+		/**
+		  * @param targetUnit Another unit
+		  * @return Modifier that needs to be applied to this unit in order to produce a value in the specified unit
+		  */
+		protected def _conversionModifierTo(targetUnit: MeterUnit) = scale.modifierTo(targetUnit.scale)
+	}
+	
+	
+	// VALUES   -------------------------
+	
+	/**
+	  * Micrometer length unit (1/1000/1000 of a meter)
+	  */
+	case object MicroMeter extends MeterUnit
+	{
+		override def scale: MetricScale = Micro
+		
+		override def conversionModifierTo(targetUnit: DistanceUnit): Double = targetUnit match {
+			case MicroMeter => 1.0
+			case u: MeterUnit => _conversionModifierTo(u)
+			case _ => toCm * CentiMeter.conversionModifierTo(targetUnit)
+		}
+	}
 	/**
 	 * Millimeter length unit (1/1000 of a meter)
 	 */
-	case object MilliMeter extends DistanceUnit
+	case object MilliMeter extends MeterUnit
 	{
-		override val abbreviation: String = "mm"
+		override def scale: MetricScale = Milli
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit): Double = targetUnit match {
 			case MilliMeter => 1.0
-			case CentiMeter => 0.1
-			case Meter => 0.001
-			case _ => toCm * CentiMeter.conversionModifierFor(targetUnit)
+			case u: MeterUnit => _conversionModifierTo(u)
+			case _ => toCm * CentiMeter.conversionModifierTo(targetUnit)
 		}
 	}
 	/**
 	 * Centimeter length unit (1/100 of a meter)
 	 */
-	case object CentiMeter extends DistanceUnit
+	case object CentiMeter extends MeterUnit
 	{
-		override val abbreviation: String = "cm"
+		override def scale: MetricScale = Centi
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit): Double = targetUnit match {
 			case CentiMeter => 1.0
-			case MilliMeter => 10.0
-			case Meter => 0.01
-			case KiloMeter | NauticalMile => toM * Meter.conversionModifierFor(KiloMeter)
+			case u: MeterUnit => _conversionModifierTo(u)
+			case NauticalMile => toM * Meter.conversionModifierTo(KiloMeter)
 			case Inch => 1 / 2.54
-			case _ => toInch * Inch.conversionModifierFor(targetUnit)
+			case _ => toInch * Inch.conversionModifierTo(targetUnit)
+		}
+	}
+	/**
+	  * Decimeter length unit (1/10 of a meter)
+	  */
+	case object DeciMeter extends MeterUnit
+	{
+		override def scale: MetricScale = Deci
+		
+		override def conversionModifierTo(targetUnit: DistanceUnit): Double = targetUnit match {
+			case DeciMeter => 1.0
+			case u: MeterUnit => _conversionModifierTo(u)
+			case _ => toCm * CentiMeter.conversionModifierTo(targetUnit)
 		}
 	}
 	/**
 	 * Meter length unit (standard metric length)
 	 */
-	case object Meter extends DistanceUnit
+	case object Meter extends MeterUnit
 	{
-		override val abbreviation: String = "m"
+		override def scale: MetricScale = Default
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case Meter => 1.0
-			case MilliMeter => 1000
-			case CentiMeter => 100
-			case KiloMeter => 0.001
+			case u: MeterUnit => _conversionModifierTo(u)
 			case Feet => 3.2808399
-			case NauticalMile => toKm * KiloMeter.conversionModifierFor(NauticalMile)
-			case _ => toFeet * Feet.conversionModifierFor(targetUnit)
+			case NauticalMile => toKm * KiloMeter.conversionModifierTo(NauticalMile)
+			case _ => toFeet * Feet.conversionModifierTo(targetUnit)
 		}
 	}
 	/**
 	 * Kilometer length unit (1000 meters)
 	 */
-	case object KiloMeter extends DistanceUnit
+	case object KiloMeter extends MeterUnit
 	{
-		override val abbreviation: String = "km"
+		override def scale: MetricScale = Kilo
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case KiloMeter => 1.0
-			case Meter => 1000.0
+			case u: MeterUnit => _conversionModifierTo(u)
 			case Mile => 0.621371192
 			case NauticalMile => 0.539956803
-			case _ => toM * Meter.conversionModifierFor(targetUnit)
+			case _ => toM * Meter.conversionModifierTo(targetUnit)
+		}
+	}
+	/**
+	  * Kilometer length unit (1 000 000 meters)
+	  */
+	case object MegaMeter extends MeterUnit
+	{
+		override def scale: MetricScale = Mega
+		
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
+			case MegaMeter => 1.0
+			case u: MeterUnit => _conversionModifierTo(u)
+			case _ => toKm * KiloMeter.conversionModifierTo(targetUnit)
 		}
 	}
 	
@@ -137,12 +231,12 @@ object DistanceUnit
 	{
 		override val abbreviation: String = "in"
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case Inch => 1.0
 			case Feet => 0.0833333333
-			case Mile => toFeet * Feet.conversionModifierFor(targetUnit)
+			case Mile => toFeet * Feet.conversionModifierTo(targetUnit)
 			case CentiMeter => 2.54
-			case _ => toCm * CentiMeter.conversionModifierFor(targetUnit)
+			case _ => toCm * CentiMeter.conversionModifierTo(targetUnit)
 		}
 	}
 	/**
@@ -152,12 +246,12 @@ object DistanceUnit
 	{
 		override val abbreviation: String = "ft"
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case Feet => 1.0
 			case Inch => 12.0
 			case Mile => 0.000189393939
 			case Meter => 0.3048
-			case _ => toM * Meter.conversionModifierFor(targetUnit)
+			case _ => toM * Meter.conversionModifierTo(targetUnit)
 		}
 	}
 	/**
@@ -167,12 +261,12 @@ object DistanceUnit
 	{
 		override def abbreviation: String = "mi"
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case Mile => 1.0
 			case Feet => 5280.0
 			case KiloMeter => 1.609344
 			case NauticalMile => 0.868976242
-			case _ => toFeet * Feet.conversionModifierFor(targetUnit)
+			case _ => toFeet * Feet.conversionModifierTo(targetUnit)
 		}
 	}
 	
@@ -183,12 +277,12 @@ object DistanceUnit
 	{
 		override val abbreviation: String = "NM"
 		
-		override def conversionModifierFor(targetUnit: DistanceUnit) = targetUnit match {
+		override def conversionModifierTo(targetUnit: DistanceUnit) = targetUnit match {
 			case NauticalMile => 1.0
 			case Meter => 1852.0
 			case KiloMeter => 1.852
 			case Mile => 1.15077945
-			case _ => toM * Meter.conversionModifierFor(targetUnit)
+			case _ => toM * Meter.conversionModifierTo(targetUnit)
 		}
 	}
 }