+++ Paradigm/test/utopia/paradigm/test/DistanceTest.scala
@@ -2,6 +2,7 @@ package utopia.paradigm.test
 
 import utopia.flow.operator.equality.EqualsExtensions._
 import utopia.paradigm.measurement.Distance
+import utopia.paradigm.measurement.MetricScale._
 
 /**
   * Tests some distance conversions
@@ -10,6 +11,17 @@ import utopia.paradigm.measurement.Distance
   */
 object DistanceTest extends App
 {
+	assert(Default.modifierTo(Deci) == 10, Default.modifierTo(Deci))
+	assert(Deci.modifierTo(Default) == 0.1)
+	assert(Deci.modifierTo(Milli) == 100)
+	assert(Milli.modifierTo(Deci) == 0.01)
+	assert(Mega.modifierTo(Kilo) == 1000)
+	
+	assert(Default.modifierFrom(Centi) == 0.01)
+	assert(Centi.modifierFrom(Default) == 100)
+	assert(Kilo.modifierFrom(Mega) == 1000)
+	assert(Mega.modifierFrom(Kilo) == 0.001)
+	
 	val d = Distance.ofMeters(154.3)
 	val d2 = Distance.ofCm(15.0)
 	val d3 = d - d2